requirements.txt
Flask==2.0.1
paho-mqtt==1.5.1
